refactor(decompose): split webapp/e2e-tests/electron/for_feature_development_not_LT_verification/graph/perf/electron-500-node-realistic-perf.spec.ts into folder
